Fill Values From Zip

Set values, warning and error messages for the import module such that success, warnings and errors are displayed inside the pop-up modal.

Usage

fillValuesFromZip(values, filename, importZip, expectedFileInZip)

Arguments

values

(reactiveValues) empty list of values in the format of the output of loadDataWrapper

filename

(character) name of the loaded file

importZip

(list) output of loadModel()

expectedFileInZip

(character) (optional) This parameter is ignored if importType != "zip". File names that must be contained in the zip upload.

Value

(list) list of values in the format of the output of loadDataWrapper