Server function for data import

Backend for data import module

Usage

importDataServer(
  id,
  title = "",
  defaultSource = c("ckan", "file", "url", "remoteModel"),
  ckanFileTypes = c("xls", "xlsx", "csv", "odt", "txt"),
  ignoreWarnings = FALSE,
  importType = c("data", "model", "zip"),
  rowNames = reactiveVal(NULL),
  colNames = reactiveVal(NULL),
  customWarningChecks = list(),
  customErrorChecks = list(),
  batch = FALSE,
  outputAsMatrix = FALSE,
  fileExtension = "zip",
  mainFolder = NULL,
  subFolder = NULL,
  rPackageName = "",
  onlySettings = FALSE,
  expectedFileInZip = c(),
  options = importOptions()
)

Arguments

id

namespace id

title

title of data import module

defaultSource

(character) default source for input "Source", e.g. "ckan", "file", or "url"

ckanFileTypes

(character) file types allowed for import from Pandora ("ckan")

ignoreWarnings

TRUE to enable imports in case of warnings

importType

(character) type of import, either "data" or "model" or "zip". ImportType == "zip" enables the optional parameter 'expectedFileInZip'.

rowNames

(reactive) use this for rownames of imported data. This parameter is ignored if importType == "model"

colNames

(reactive) use this for colnames of imported data. This parameter is ignored if importType == "model"

customWarningChecks

list of reactive(!) functions which will be executed after importing of data. functions need to return TRUE if check is successful or a character with a warning otherwise. This parameter is ignored if importType == "model"

customErrorChecks

list of reactive(!) functions which will be executed after importing of data. functions need to return TRUE if check is successful or a character with a warning otherwise. This parameter is ignored if importType == "model"

batch

(logical) use batch import. This parameter is ignored if importType == "model"

outputAsMatrix

(logical) TRUE if output must be a matrix, e.g. for batch = TRUE in Resources. This parameter is ignored if importType == "model"

fileExtension

(character) (otional) app specific file extension, e.g. "resources", "bmsc", "bpred", or (app-unspecific) "zip". Only files with this extension are valid for import.

mainFolder

(character) DEPRECATED. folder containing all loadable .zip files. This parameter is ignored if importType == "data"

subFolder

(character) (optional) subfolder containing loadable .zip files. This parameter is ignored if importType == "data"

rPackageName

(character) DEPRECATED. Instead, please use options = importOptions(rPackageName = <your package>).

onlySettings

(logical) if TRUE allow only upload of user inputs and user data. This parameter is ignored if importType == "data"

expectedFileInZip

(character) (optional) This parameter is ignored if importType != "zip". File names that must be contained in the zip upload.

options

(list) Extra options for the import module.