Load Model Wrapper

Wrapper function to load a model from a zip file. File extension may differ from app to app, but in essence all models are stored in a zip file. The zip file contains a model object, a data object, an inputs object and potentially help files. This function unzips the file, extracts the model object, and checks if the model is valid for the app. It returns a list with the model object, the data object, the inputs object and gives a message if the model was successfully loaded.

Usage

loadModelWrapper(
  values,
  filepath,
  filename,
  subFolder,
  rPackageName,
  onlySettings,
  fileExtension = "zip",
  ...
)

Arguments

values

(list) list with import specifications

filepath

(character) path to the model file

filename

(character) name of the model file

subFolder

(character) (optional) subfolder containing loadable .zip files

rPackageName

(character) name of the package (as in the Description file) in which this module is called. This value is needed to determine the Github repository of the package for loading example models or zip or data links.

onlySettings

(logical) if TRUE allow only upload of user inputs and user data

fileExtension

(character) (optional) app specific file extension, e.g. "resources", "bpred", "bmsc"

...

parameters for other wrappers