Select Data Server

Server function of the module

Usage

selectDataServer(
  id,
  importType = "data",
  mergeList,
  customNames,
  dataSource,
  subFolder = NULL,
  ignoreWarnings = FALSE,
  rPackageName = "",
  onlySettings = FALSE,
  fileExtension = "zip",
  expectedFileInZip = c()
)

Arguments

id

id of module

importType

(character) type of import, either "data", "model", "zip" or "list". ImportType == "model" expects a zip file containing a model. The file will be unzipped, the model object extracted, and checked if it is valid for the app. ImportType == "zip" enables the optional parameter 'expectedFileInZip'. The file is validated and the path to the zip file will be returned.

mergeList

(reactiveVal) list of data imports submitted for data processing via buttons 'Create Query with data' or 'Prepare / Merge data'

customNames

settings for custom column and row names

dataSource

(reactiveValues) path, filename, type and input, output of selectSourceServer()

subFolder

(character) (optional) subfolder containing loadable .zip files. This parameter is ignored if importType == "data"

ignoreWarnings

TRUE to enable imports in case of warnings

rPackageName

(character) DEPRECATED. Instead, please use options = importOptions(rPackageName = <your package>).

onlySettings

(logical) if TRUE allow only upload of user inputs and user data. This parameter is ignored if importType == "data"

fileExtension

(character) (otional) app specific file extension, e.g. "resources", "bmsc", "bpred", or (app-unspecific) "zip". Only files with this extension are valid for import.

expectedFileInZip

(character) (optional) This parameter is ignored if importType != "zip". File names that must be contained in the zip upload.